deleted all other server and client files. Added 10 tests (7 for server, 3 for client). All functionality works
