+++ client/client.py
@@ -1,36 +1,156 @@
+"""
+Reference: https://www.rabbitmq.com/tutorials/tutorial-six-python 
+"""
+import pika
 import json
+import uuid
+import time
+from datetime import datetime
 import random
 import threading
-import time
-import uuid
-from datetime import datetime
 
 import customtkinter as ctk
-import pika
+
+# To run rabbit mq:
+# 1. ran the cmd: docker-compose up
+# 2. login using the following credentials guest/guest on the port http://localhost:15672
+
+class SmartMeterClient:
+    def __init__(self, app):
+        self.connection = None
+        self.channel = None
+        self.response = None
+        self.corr_id = None
+        self.reading = 0.0
+        self.reading_id = str(uuid.uuid4())
+        self.user_email = str(uuid.uuid4()) + "@gmail.com"
+        # update UI
+        self.app = app
+        self.connected = False
+
+        # check connection to the server initially and periodically
+        self.connect_to_server()
+        connection_thread = threading.Thread(target=self.check_connection_status, daemon=True)
+        connection_thread.start() 
+
+    def connect_to_server(self):
+        try:
+            self.connection = pika.BlockingConnection(pika.ConnectionParameters('localhost'))
+            self.channel = self.connection.channel()
+
+            self.channel.basic_consume(
+                queue="amq.rabbitmq.reply-to",
+                on_message_callback=self.on_response,
+                auto_ack=True
+            )
+
+            self.connected = True
+            self.app.update_connection_status("connected")
+
+        except pika.exceptions.AMQPConnectionError as e:
+            print(f"Error connecting to RabbitMQ: {e}")
+            self.app.update_connection_status("error")
+
+    def check_connection_status(self):
+        while True:
+            # if connection is set to none or connection has been closed then status changed to error
+            if self.connection is None or self.connection.is_closed:
+                self.connected = False
+                self.app.update_connection_status("error")
+                # and try to reconnect
+                self.connect_to_server()
+            else:
+                self.connected = True
+                
+            time.sleep(5)
+
+    def close(self):
+        if self.is_connected():
+            self.connection.close()
+            self.connection = None
+            print("Closed RabbitMQ connection")
+
+    def on_response(self, ch, method, props, body):
+            if self.corr_id == props.correlation_id:
+                self.response = body
+
+    def generate_meter_reading(self):
+        return {
+            "id": self.reading_id,
+            "user_email": self.user_email,
+            "meter_reading": self.reading,
+            "timestamp": datetime.now().isoformat(),
+        }
+
+    #sends meter reading
+    def send_meter_reading(self, reading_data):
+        self.response = None
+        self.corr_id = str(uuid.uuid4())
+
+        #publishes the rpc request to the server
+        self.channel.basic_publish(
+            exchange='',
+            routing_key='rpc_queue',
+            properties=pika.BasicProperties(
+                reply_to="amq.rabbitmq.reply-to",
+                correlation_id=self.corr_id,
+            ),
+            body=json.dumps(reading_data)
+        )   
+
+        # timeout if no response from server
+        start_time = time.time()
+        timeout = 10
+
+        # waits for the response
+        while self.response is None:
+            self.connection.process_data_events()
+            # if the difference between now time and starting time is greater than timeout (10 sec) it will return none and loop out
+            if time.time() - start_time > timeout:
+                print("Time out: No response!")
+                return None
+        return self.response      
+    
+    def update_meter_readings(self, app):
+        # random increase in reading
+        self.reading += random.uniform(1.0, 5.0)
+
+        reading_data = self.generate_meter_reading()
+        print(f"Meter reading sent: {reading_data} kWh")
+        updated_price = self.send_meter_reading(reading_data)
+
+        #decode from bytes to string and then float for correct format
+        updated_price = updated_price.decode()  
+        updated_price = float(updated_price) 
+
+        print(f"Updated Price: £{updated_price:.2f}")
+        app.update_main_display(f"£{updated_price:.2f}", f"{self.reading:.2f} kWh")
+
+    def start_meter_updater(self, app):
+        while True:
+            wait_time = random.randint(15, 60)
+            print(f"Waiting {wait_time} seconds to send...")
+            time.sleep(wait_time)
+            self.update_meter_readings(app)
 
 
 class SmartMeterUI(ctk.CTk):
-    """
-    TODO: Add docstrings to classes and methods
-    """
-
     def __init__(self):
         super().__init__()
 
         # Appearance and theme
         self.title("Smart Meter Interface")
         self.geometry("600x300")
-        # Use native system dark / light mode
-        ctk.set_appearance_mode("system")
+        ctk.set_appearance_mode("dark")
         ctk.set_default_color_theme("blue")
 
         # Configure Grid layout
-        self.grid_columnconfigure(0, weight=1)
+        self.grid_columnconfigure(0, weight=1)  # Create a single column
         self.grid_rowconfigure(0, weight=1)  # Server Connection status section
         self.grid_rowconfigure(1, weight=3)  # Main display
         self.grid_rowconfigure(2, weight=1)  # Outage Notice section
 
-        # Define attributes
+        # UI elements initialization
         self.connection_status_frame = None
         self.connection_status_label = None
         self.time_label = None
@@ -57,6 +177,7 @@ class SmartMeterUI(ctk.CTk):
             row=0, column=0, padx=20, pady=10, sticky="nsew"
         )
 
+        # Time label
         self.time_label = ctk.CTkLabel(
             self.connection_status_frame,
             text="11:11",
@@ -65,6 +186,7 @@ class SmartMeterUI(ctk.CTk):
         )
         self.time_label.pack(side="top", anchor="ne", padx=10)
 
+        # Connection status label
         self.connection_status_label = ctk.CTkLabel(
             self.connection_status_frame,
             text="Connecting to server...",
@@ -77,13 +199,13 @@ class SmartMeterUI(ctk.CTk):
         self.middle_frame = ctk.CTkFrame(self, corner_radius=10, fg_color="grey12")
         self.middle_frame.grid(row=1, column=0, padx=20, pady=10, sticky="nsew")
 
-        self.price_label = ctk.CTkLabel(
+        self.amount_label = ctk.CTkLabel(
             self.middle_frame,
             text="£x.xx",
             font=("Arial", 36, "bold"),
             text_color="green",
         )
-        self.price_label.pack(pady=(20, 5))
+        self.amount_label.pack(pady=(20, 5))
 
         self.usage_label = ctk.CTkLabel(
             self.middle_frame,
@@ -116,8 +238,8 @@ class SmartMeterUI(ctk.CTk):
                 text_color="gold",
             )
 
-    def update_main_display(self, price, usage):
-        self.price_label.configure(text=f"{price}")
+    def update_main_display(self, amount, usage):
+        self.amount_label.configure(text=amount)
         self.usage_label.configure(text=f"Used so far today: {usage}")
 
     def update_notice_message(self, message):
@@ -130,146 +252,16 @@ class SmartMeterUI(ctk.CTk):
     def update_time(self):
         current_time = datetime.now().strftime("%H:%M")
         self.time_label.configure(text=current_time)  # Update the time label
-        self.after(1000, self.update_time)  # Set to update every second
+        self.after(1000, self.update_time)  # Update every second
 
 
-# TODO: Could do with decoupling this from the UI, perhaps by using a
-# controller class?
-class SmartMeterClient:
-    def __init__(self):
-        # TODO: Add error handling for connection failure, + retry logic (on another thread)
-        # into SelectConnection instead of BlockingConnection
-        self.connection = pika.BlockingConnection(
-            pika.ConnectionParameters("localhost")
-        )
-        self.channel = self.connection.channel()
-
-        # Set up a listener for direct reply-to using RPC (no need to declare a
-        # queue)
-        self.channel.basic_consume(
-            queue="amq.rabbitmq.reply-to",
-            on_message_callback=self.on_response,
-            auto_ack=True,
-        )
-
-        self.response = None
-        self.corr_id = None
-        self.reading = 0.0  # TODO: Initial reading, need to populate this with a user's reading after user auth is implemented
-        self.timeout_reading = None
-
-    def on_response(self, _ch, _method, properties, body):
-        if self.corr_id == properties.correlation_id:
-            self.response = body
-
-    @staticmethod
-    def generate_meter_reading(reading_id, reading):
-        return {
-            "id": reading_id,
-            "user_email": "shu@example.com",  # TODO: Replace with the user's email or ID
-            "meter_reading": reading,
-            "timestamp": datetime.now().isoformat(),
-        }
-
-    def send_meter_reading(self, reading_data):
-        self.response = None
-        self.corr_id = str(uuid.uuid4())  # Create a unique correlation ID
-
-        # Publish the meter reading to the server
-        self.channel.basic_publish(
-            exchange="",
-            routing_key="meter_reading_queue",  # Queue where the server listens
-            properties=pika.BasicProperties(
-                reply_to="amq.rabbitmq.reply-to",  # Use direct reply-to
-                correlation_id=self.corr_id,  # Include the correlation ID
-            ),
-            body=json.dumps(reading_data),
-        )  # TODO: Error handling
-
-        start_time = time.time()  # Record the start time
-        # Set the timeout duration in seconds, must be less than MIN_WAIT
-        TIMEOUT_DURATION = 10
-
-        # * Wait for the response, blocking until the response is received or a timeout, so must be ran in a separate thread
-        while self.response is None:
-            self.connection.process_data_events()
-            if time.time() - start_time > TIMEOUT_DURATION:
-                return "timeout"
-
-        return self.response.decode("utf-8")  # Return the response as a string
-
-    def update_meter_readings(self, app):
-        MIN_INCREASE = 1.0  # TODO: Should these be moved to class attributes?
-        MAX_INCREASE = 5.0
-        # Generate a random increase in the reading
-        increase = random.uniform(MIN_INCREASE, MAX_INCREASE)
-        self.reading += increase  # Update the reading
-
-        # Send the meter reading to the server and receiving an updated price
-        reading_data = self.generate_meter_reading(1, self.reading)
-        updated_price = self.send_meter_reading(reading_data)
-        print(f"Price: {updated_price}")
-        if self.timeout_reading and updated_price == "timeout":
-            price_text = app.price_label.cget("text")
-            app.update_main_display(price_text, f"{self.reading:.2f} kWh")
-        elif updated_price == "timeout":
-            app.update_connection_status("error")
-            # show last know price @ last known reading
-            original_price = app.price_label.cget("text")
-            # part of logic to show last known price/reading
-            self.timeout_reading = self.reading - increase
-            if original_price in ("£x.xx", "£?.??"):
-                app.update_main_display("£?.??", f"{self.reading:.2f} kWh")
-            else:
-                app.update_main_display(
-                    f"£?.?? - was {original_price} @ {self.timeout_reading:.2f} kWh",
-                    f"{self.reading:.2f} kWh",
-                )
-
-        else:
-            self.timeout_reading = None
-            app.update_connection_status("connected")
-            # Convert the price to float and format to two decimal places
-            updated_price = (
-                f"{float(updated_price):.2f}"  # Format to two decimal places
-            )
-
-            # Update the main display
-            app.update_main_display(f"£{updated_price}", f"{self.reading:.2f} kWh")
-
-    def start_meter_updater(self, app):
-        MIN_WAIT = 15  # Must be greater than TIMEOUT_DURATION
-        MAX_WAIT = 60
-
-        while True:
-            wait_time = random.randint(MIN_WAIT, MAX_WAIT)
-            print(f"Waiting for {wait_time} seconds...")
-            time.sleep(wait_time)
-            print("Done waiting")
-            reading_thread = threading.Thread(
-                target=self.update_meter_readings, args=(app,), daemon=True
-            )
-            reading_thread.start()
-
-
-# Main app with mock data
 if __name__ == "__main__":
     app = SmartMeterUI()
 
-    # TODO: Simulate notice message, implement notice system
-    app.after(
-        5000,
-        lambda: app.update_notice_message(
-            "Alert: Possible electricity grid problem detected."
-        ),
-    )
+    client = SmartMeterClient(app)
 
-    # Usage
-    client = SmartMeterClient()
-
-    # Start the meter reading updates in a separate thread
-    reading_thread = threading.Thread(
-        target=client.start_meter_updater, args=(app,), daemon=True
-    )
+    #start meter reading updates in a separate thread
+    reading_thread = threading.Thread(target=client.start_meter_updater, args=(app,), daemon=True)
     reading_thread.start()
 
     app.mainloop()
